Remove apollo server shutdown method
The shutdown() method was leading to tracbacks when shutting down the
Apollo server because the version of the temporalio module we are using
doesn't have the previously specified close() method. Because the only
action the shutdown() method took was to ensure the temporal connection
was closed and that is managed automatically by this version of
temporalio the entire method was removed.