Refactor supported_product_mirror pages
This is a refactor of the supported_product_mirror related pages that is
focused on removing in-line and custom styling outside of the Carbon
Design System. While doing the refactor of
admin_supported_product_mirror.jinja I found the JavaScript
implimentation for presenting tabs with override and adivsory blocks
information wasn't going to be easy to replace with native Carbon Design
System elements. I also realized that trying to display all of the
blocks in a tab wasn't going to scale when there are thounsands of
blocks. This led to the decision to break the blocks and overrides into
distinct pages that would function more like the Advisories pages where
a user would be shown a discrete number at a given time and would have
the ability to search for the appropriate advisory.